add validations

+++ db/migrate/20201016163810_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration[6.0]
   def change
     create_table :products do |t|
       t.column(:name, :string)
-      t.column(:cost, :decimal)
+      t.column(:cost, :integer)
       t.column(:country_of_origin, :string)
       
       t.timestamps()
